Add html, Not finished

+++ blog/views.py
@@ -1,9 +1,17 @@
 from django.shortcuts import render
 from django.utils import timezone
-from .models import Post, Comment
-from .forms import PostForm, CommentForm
 from django.shortcuts import render, get_object_or_404, redirect
 from django.contrib.auth.decorators import login_required
+
+from .models import Post, Comment
+from .forms import PostForm, CommentForm
+
+from django.contrib.auth import login, authenticate
+from django.contrib.auth.forms import UserCreationForm
+from django.shortcuts import render, redirect
+
+def main (request) :
+    return render(request, 'blog/main.html')
 
 def post_list(request) :
     posts = Post.objects.filter(published_date__lte=timezone.now()).order_by('published_date')
@@ -29,6 +37,9 @@ def post_new (request):
 @login_required
 def post_edit (request, pk):
     post = get_object_or_404(Post, pk=pk)
+    if (post.author != request.user) :
+        return redirect('post_detail', pk=post.pk)
+    
     if (request.method == "POST") :
         form = PostForm(request.POST, instance=post)
         if form.is_valid() :
@@ -42,7 +53,7 @@ def post_edit (request, pk):
 
 @login_required
 def post_draft_list(request):
-    posts = Post.objects.filter(published_date__isnull=True).order_by('created_date')
+    posts = Post.objects.filter(published_date__isnull=True, author=request.user).order_by('created_date')
     return render(request, 'blog/post_draft_list.html', {'posts': posts})
 
 def publish(self):
@@ -58,6 +69,10 @@ def post_publish(request, pk):
 @login_required
 def post_remove(request, pk):
     post = get_object_or_404(Post, pk=pk)
+
+    if (post.author != request.user) :
+        return redirect('post_detail', pk=post.pk)
+    
     post.delete()
     return redirect('post_list')
 
@@ -78,6 +93,9 @@ def add_comment_to_post(request, pk):
 @login_required
 def comment_approve(request, pk):
     comment = get_object_or_404(Comment, pk=pk)
+    if (comment.author != request.user) :
+        return redirect('post_detail', pk=comment.post.pk)
+    
     comment.approve()
     return redirect('post_detail', pk=comment.post.pk)
 
@@ -86,3 +104,17 @@ def comment_remove(request, pk):
     comment = get_object_or_404(Comment, pk=pk)
     comment.delete()
     return redirect('post_detail', pk=comment.post.pk)
+
+def signup(request):
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            form.save()
+            username = form.cleaned_data.get('username')
+            raw_password = form.cleaned_data.get('password1')
+            user = authenticate(username=username, password=raw_password)
+            login(request, user)
+            return redirect('main')
+    else:
+        form = UserCreationForm()
+    return render(request, 'blog/signup.html', {'form': form})
